+++ examples/gjs/gtk-3-application/main.ts
@@ -3,7 +3,7 @@
 // SPDX-FileCopyrightText: 2017 Andy Holmes <andrew.g.r.holmes@gmail.com>
 // Based on https://gitlab.gnome.org/GNOME/gjs/-/blob/master/examples/gtk-application.js
 
-import './@types/gjs/index.js';
+import './@types/index.js';
 
 // Include this in case both GTK3 and GTK4 installed, otherwise an exception
 // will be thrown