+++ examples/gjs/gtk-3-calc/main.ts
@@ -3,7 +3,7 @@
 // SPDX-FileCopyrightText: 2008 Robert Carr <carrr@rpi.edu>
 // Based on https://gitlab.gnome.org/GNOME/gjs/-/blob/master/examples/calc.js
 
-import './@types/gjs/index.js';
+import './@types/index.js';
 import Gtk from 'gi://Gtk?version=3.0';
 
 Gtk.init(null);