+++ examples/gjs/gtk-3-editor/main.ts
@@ -1,7 +1,7 @@
-import './@types/gjs/index'
-import * as Gtk from './@types/gjs/gtk-3.0'
-import * as GtkSource from './@types/gjs/gtksource-3.0'
-import { gettext } from './@types/gjs/gjs'
+import './@types/index'
+import * as Gtk from './@types/gtk-3.0'
+import * as GtkSource from './@types/gtksource-3.0'
+import { gettext } from './@types/gjs'
 
 Gtk.init(null)