+++ examples/gjs/gtk-3-template/main.ts
@@ -3,7 +3,7 @@
 // SPDX-FileCopyrightText: 2021 Andy Holmes <andyholmes@gnome.org>
 // Based on https://gitlab.gnome.org/GNOME/gjs/-/blob/master/examples/gtk3-template.js
 
-import './@types/gjs/index.js';
+import './@types/index.js';
 
 import GObject from 'gi://GObject?version=2.0';
 import Gio from 'gi://Gio?version=2.0';