+++ examples/gjs/gtk-3-webkit/main.ts
@@ -3,7 +3,7 @@
 // SPDX-FileCopyrightText: 2008 litl, LLC
 // Based on https://gitlab.gnome.org/GNOME/gjs/-/blob/master/examples/webkit.js
 
-import './@types/gjs/index.js';
+import './@types/index.js';
 
 import Gtk from 'gi://Gtk?version=3.0';
 import WebKit from 'gi://WebKit2?version=4.0';