+++ examples/gjs/soup-3-http/http-server.ts
@@ -5,7 +5,7 @@
 // This is a simple example of a HTTP server in Gjs using libsoup 3
 // https://gitlab.gnome.org/GNOME/gjs/-/blob/master/examples/http-server.js
 
-import './@types/gjs/index.js';
+import './@types/index.js';
 import GLib from 'gi://GLib?version=2.0';
 import Soup from 'gi://Soup?version=3.0';