+++ examples/gjs/soup-3-websocket/client.ts
@@ -7,12 +7,12 @@
 
 // Based on https://gitlab.gnome.org/GNOME/gjs/-/blob/master/examples/websocket-client.js
 
-import './@types/gjs/index.js';
+import './@types/index.js';
 
 import GLib from 'gi://GLib?version=2.0';
 import Soup from 'gi://Soup?version=3.0';
 import Gio from 'gi://Gio?version=2.0';
-import { byteArray } from'./@types/gjs/gjs'
+import { byteArray } from'./@types/gjs'
 
 const loop = GLib.MainLoop.new(null, false);