+++ examples/node-gtk/soup-3-http/http-client.ts
@@ -6,10 +6,10 @@
 // https://developer.gnome.org/libsoup/stable/libsoup-client-howto.html
 // https://gitlab.gnome.org/GNOME/gjs/-/blob/master/examples/http-client.js
 
-import './@types/node-gtk/index.js';
-import GLib from './@types/node-gtk/node-glib-2.0.js';
-import Soup from './@types/node-gtk/node-soup-3.0.js';
-import Gio from './@types/node-gtk/node-gio-2.0.js';
+import './@types/index.js';
+import GLib from './@types/node-glib-2.0.js';
+import Soup from './@types/node-soup-3.0.js';
+import Gio from './@types/node-gio-2.0.js';
 
 const loop = GLib.MainLoop.new(null, false);