+++ NEWS.md
@@ -2,6 +2,7 @@
 - New test: Create simple vala classes, build the gir files from that, generate the typescript type definitions for that and test if the result looks good
 - Create a Reporter to create a text or json file with warnings like unknown types, renaming, etc and a summary of all that (e.g. 5 type conflicts resolved, 8 unknown types of *gint)
 - Update types for Gjs v1.75.2, see https://gitlab.gnome.org/GNOME/gjs/-/commit/666755b3b09d765e43d415e76105b828517b5509
+- Make use of packages/lib/templates/gjs/types.d.ts
 
 # Unreleased
 - Add support to generate a NPM package for each generated module type definition